+++ packages/transfer_manager_android/lib/transfer_manager_android.dart
@@ -1,5 +1,6 @@
 library;
 
 export 'src/background_download_engine.dart';
+export 'src/background_tus_upload_engine.dart';
 export 'src/background_upload_engine.dart';
 export 'src/method_channel_platform.dart';